Refactor BookList component to use custom hooks
Extraí a funcionalidade de busca de livros para um hook personalizado, melhorando a reutilização de código e o gerenciamento de efeitos colaterais, e introduzi um hook de debounce para otimizar o desempenho da filtragem. Além disso, criei um novo componente BookCard para melhorar a modularidade e a legibilidade.
